examples/01-basic-examples/hello-world/requirements.txt
# TFrameX Hello World Example Dependencies

# Core TFrameX framework
tframex>=0.1.3

# Required for environment configuration
python-dotenv>=1.0.0

# Optional: Enhanced logging (if needed)
colorama>=0.4.6